requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
dnspython==2.7.0
pymongo==4.14.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
